Add triggers and alarms for Neutron agents
This patch adds monitoring of the neutron-data cluster when Neutron uses
OVS instead of OpenContrail.

It adds new triggers for l3, metadata and openvswitch agents to detect
whether one agent is down, only 2 agents are still up or all are down.
It also adds a new AFD based on these triggers and its corresponding GSE
on the aggregator.

Change-Id: I71a4f87b66d4ef1c44efd394c2369aceed90098e