Remove the usage of down and disabled together
This patch removes the trigger with state '== down || == disabled'
because it is not working. When stopping 2 L3 instances over 3 we see
the severity OKAY. So there is an issue here. To fix this we use the
single test 'up' and we ensure that it triggers only if one agent is up
and at least another one is down or disabled.

Change-Id: I4e64e4302bd4780401fb1c90b1dc8eeb3d3f9e05